requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
beautifulsoup4==4.9.2
black==20.8b1
bs4==0.0.1
cachelib==0.1.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
docopt==0.6.2
Flask==1.1.2
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==2.4.4
google-api-core==1.22.2
google-api-python-client==1.12.2
google-auth==1.21.3
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
mysqlclient==2.0.1
oauth2client==4.1.3
oauthlib==2.1.0
pathspec==0.8.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==0.10.1
python-dotenv==0.14.0
pytz==2020.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.1.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.19
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
yarg==0.1.9
